Refactored into two tests
